Vuex: Create separate subfolder for store and states, remove extra constant with module name
