--- src/constants/vuex.ts
@@ -1,1 +1,0 @@
-export const FilmsModuleName: string = 'films';
