+++ src/vuex/store/index.ts
@@ -1,7 +1,6 @@
-import Films from './modules/films/state';
+import Films, { FilmsModuleName } from '../modules/films/state';
 import Vue from 'vue';
 import Vuex from 'vuex';
-import { FilmsModuleName } from '@/constants/vuex';
 
 Vue.use(Vuex);
